+++ src/radiator/ui/pipeline.clj
@@ -28,6 +28,24 @@
     (some #(= :in-progress (:pipeline-status %)) pipelines) img/codepipeline-ongoing
     :else img/codepipeline-succeed))
 
+(defn first-line-of-commit
+  [pipeline]
+  (when-let [message (:message pipeline)]
+    (-> message
+        clojure.string/split-lines
+        first)))
+
+(defn add-author-if-present
+  [pipeline]
+  (when-let [author (:author pipeline)]
+    (str ": (" (:author pipeline) ")")))
+
+(defn commit-info
+  [pipeline]
+  [:div.container-column [:p.small [:strong (:name pipeline)] (add-author-if-present pipeline)]
+   (if-let [first-line (first-line-of-commit pipeline)]
+     [:p.small first-line])])
+
 (defn pipelines-box
   [combined-pipelines]
   [:div.light-grey-background
@@ -43,7 +61,7 @@
      (for [pipeline (filter-ok combined-pipelines)]
        [:div.t-b-5px-padding
         (cond
-          (= (:pipeline-status pipeline) :success) [:div img/codepipeline-succeed [:span (:name pipeline)]]
-          (= (:pipeline-status pipeline) :in-progress) [:div img/codepipeline-ongoing [:span (:name pipeline)]]
-          (= (:pipeline-status pipeline) :unknown) [:div img/codepipeline-ongoing [:span (str "Unknown: " (:name pipeline))]]
-          :else [:div img/alarm-30px [:span (:name pipeline)]])]))])
+          (= (:pipeline-status pipeline) :success) [:div.flex img/codepipeline-succeed (commit-info pipeline)]
+          (= (:pipeline-status pipeline) :in-progress) [:div.flex img/codepipeline-ongoing (commit-info pipeline)]
+          (= (:pipeline-status pipeline) :unknown) [:div.flex img/codepipeline-ongoing (commit-info pipeline)]
+          :else [:div.flex img/alarm-30px (commit-info pipeline)])]))])