+++ ChangeLog.md
@@ -1,3 +1,7 @@
+## 2018-01-14
+* Show commit author and message if known in pipeline status. Requires 
+Radiator Exposer update also.
+
 ## 2018-01-10
 * Show alarms which have happened in the last 24 hours. Requires 
 Radiator Exposer update also.